Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/pixel-manipulation/2d.imageData.get.zero-expected.txt

2d.imageData.get.zero

getImageData() throws INDEX_SIZE_ERR if size is zero


PASS getImageData() throws INDEX_SIZE_ERR if size is zero
